--- a/frontend/src/components/mobile/MobileDebugger.css
+++ b/frontend/src/components/mobile/MobileDebugger.css
@@ -1,0 +1,412 @@
+.debug-floating-button {
+  position: fixed;
+  bottom: 20px;
+  right: 20px;
+  z-index: 10000;
+  width: 50px;
+  height: 50px;
+  border-radius: 50%;
+  background: #ff6b35;
+  color: white;
+  font-size: 20px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.3);
+  transition: all 0.3s ease;
+}
+
+.debug-floating-button:hover {
+  transform: scale(1.1);
+  box-shadow: 0 6px 16px rgba(0, 0, 0, 0.4);
+}
+
+.mobile-debugger {
+  position: fixed;
+  top: 10px;
+  left: 10px;
+  right: 10px;
+  bottom: 80px;
+  background: rgba(0, 0, 0, 0.95);
+  color: #e0e0e0;
+  border-radius: 8px;
+  z-index: 9999;
+  display: flex;
+  flex-direction: column;
+  font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
+  font-size: 12px;
+  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.5);
+  backdrop-filter: blur(10px);
+}
+
+.debugger-header {
+  padding: 1rem;
+  border-bottom: 1px solid #333;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.debugger-header h3 {
+  margin: 0;
+  color: #ff6b35;
+  font-size: 1.2rem;
+  font-weight: bold;
+}
+
+.debugger-controls {
+  display: flex;
+  gap: 0.5rem;
+  align-items: center;
+  flex-wrap: wrap;
+}
+
+.debug-filter,
+.debug-search {
+  padding: 0.4rem;
+  border: 1px solid #555;
+  background: #333;
+  color: #e0e0e0;
+  border-radius: 4px;
+  font-size: 12px;
+}
+
+.debug-filter {
+  min-width: 80px;
+}
+
+.debug-search {
+  min-width: 120px;
+}
+
+.debug-search::placeholder {
+  color: #999;
+}
+
+.clear-logs-btn,
+.export-logs-btn {
+  background: #6c757d;
+  font-size: 11px;
+  padding: 0.4rem 0.8rem;
+}
+
+.clear-logs-btn:hover {
+  background: #5a6268;
+}
+
+.export-logs-btn:hover {
+  background: #28a745;
+}
+
+.debug-utils {
+  padding: 0.5rem 1rem;
+  border-bottom: 1px solid #333;
+  display: flex;
+  gap: 0.5rem;
+  overflow-x: auto;
+}
+
+.debug-util-btn {
+  background: #495057;
+  font-size: 10px;
+  padding: 0.3rem 0.6rem;
+  white-space: nowrap;
+  flex-shrink: 0;
+}
+
+.debug-util-btn:hover {
+  background: #6c757d;
+}
+
+.logs-container {
+  flex: 1;
+  overflow-y: auto;
+  padding: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  gap: 0.25rem;
+}
+
+.no-logs {
+  text-align: center;
+  color: #999;
+  padding: 2rem;
+  font-style: italic;
+}
+
+.log-entry {
+  background: rgba(255, 255, 255, 0.05);
+  border-radius: 4px;
+  border-left: 3px solid #6c757d;
+  transition: all 0.2s ease;
+}
+
+.log-entry:hover {
+  background: rgba(255, 255, 255, 0.08);
+}
+
+.log-entry.log-error {
+  border-left-color: #dc3545;
+  background: rgba(220, 53, 69, 0.1);
+}
+
+.log-entry.log-warn {
+  border-left-color: #ffc107;
+  background: rgba(255, 193, 7, 0.1);
+}
+
+.log-entry.log-info {
+  border-left-color: #17a2b8;
+  background: rgba(23, 162, 184, 0.1);
+}
+
+.log-entry.log-debug {
+  border-left-color: #28a745;
+  background: rgba(40, 167, 69, 0.1);
+}
+
+.log-header {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  padding: 0.5rem;
+  font-size: 10px;
+  opacity: 0.8;
+}
+
+.log-icon {
+  font-size: 12px;
+}
+
+.log-time {
+  color: #999;
+  min-width: 70px;
+}
+
+.log-category {
+  background: rgba(255, 255, 255, 0.1);
+  padding: 0.1rem 0.4rem;
+  border-radius: 10px;
+  font-size: 9px;
+  text-transform: uppercase;
+  letter-spacing: 0.5px;
+}
+
+.log-level {
+  color: #999;
+  font-size: 9px;
+  margin-left: auto;
+}
+
+.log-message {
+  padding: 0 0.5rem 0.5rem;
+  line-height: 1.4;
+  word-break: break-word;
+  white-space: pre-wrap;
+}
+
+.log-data {
+  margin: 0.5rem;
+  border: 1px solid #555;
+  border-radius: 4px;
+}
+
+.log-data summary {
+  padding: 0.3rem 0.5rem;
+  background: rgba(255, 255, 255, 0.05);
+  cursor: pointer;
+  font-size: 10px;
+  border-radius: 4px 4px 0 0;
+}
+
+.log-data summary:hover {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+.log-data pre {
+  margin: 0;
+  padding: 0.5rem;
+  font-size: 10px;
+  line-height: 1.3;
+  overflow-x: auto;
+  background: rgba(0, 0, 0, 0.3);
+}
+
+.debugger-status {
+  padding: 0.5rem 1rem;
+  border-top: 1px solid #333;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  font-size: 10px;
+  color: #999;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+/* Scrollbar styling for WebKit browsers */
+.logs-container::-webkit-scrollbar {
+  width: 6px;
+}
+
+.logs-container::-webkit-scrollbar-track {
+  background: rgba(255, 255, 255, 0.1);
+  border-radius: 3px;
+}
+
+.logs-container::-webkit-scrollbar-thumb {
+  background: rgba(255, 255, 255, 0.3);
+  border-radius: 3px;
+}
+
+.logs-container::-webkit-scrollbar-thumb:hover {
+  background: rgba(255, 255, 255, 0.5);
+}
+
+/* Mobile-specific adjustments */
+@media (max-width: 768px) {
+  .mobile-debugger {
+    top: 5px;
+    left: 5px;
+    right: 5px;
+    bottom: 70px;
+    font-size: 11px;
+  }
+  
+  .debugger-header {
+    padding: 0.75rem;
+    flex-direction: column;
+    align-items: stretch;
+  }
+  
+  .debugger-controls {
+    flex-direction: column;
+    align-items: stretch;
+  }
+  
+  .debug-filter,
+  .debug-search {
+    width: 100%;
+    margin-bottom: 0.25rem;
+  }
+  
+  .debug-utils {
+    padding: 0.5rem;
+  }
+  
+  .debug-util-btn {
+    font-size: 9px;
+    padding: 0.4rem 0.6rem;
+  }
+  
+  .log-header {
+    flex-wrap: wrap;
+    gap: 0.25rem;
+  }
+  
+  .log-time {
+    min-width: auto;
+  }
+  
+  .debugger-status {
+    padding: 0.5rem;
+    flex-direction: column;
+    align-items: stretch;
+    gap: 0.25rem;
+  }
+}
+
+@media (max-width: 480px) {
+  .debug-floating-button {
+    bottom: 15px;
+    right: 15px;
+    width: 45px;
+    height: 45px;
+    font-size: 18px;
+  }
+  
+  .mobile-debugger {
+    font-size: 10px;
+  }
+  
+  .debugger-header h3 {
+    font-size: 1.1rem;
+  }
+  
+  .debug-filter,
+  .debug-search {
+    font-size: 11px;
+    padding: 0.5rem;
+  }
+  
+  .log-message {
+    font-size: 11px;
+  }
+  
+  .log-data pre {
+    font-size: 9px;
+  }
+}
+
+/* Dark mode enhancements */
+@media (prefers-color-scheme: dark) {
+  .debug-floating-button {
+    background: #ff6b35;
+  }
+  
+  .mobile-debugger {
+    background: rgba(0, 0, 0, 0.98);
+    border: 1px solid #333;
+  }
+  
+  .debug-filter,
+  .debug-search {
+    background: #1a1a1a;
+    border-color: #444;
+    color: #e0e0e0;
+  }
+}
+
+/* Accessibility improvements */
+.debug-floating-button:focus,
+.clear-logs-btn:focus,
+.export-logs-btn:focus,
+.debug-util-btn:focus,
+.debug-filter:focus,
+.debug-search:focus {
+  outline: 2px solid #ff6b35;
+  outline-offset: 2px;
+}
+
+/* High contrast mode */
+@media (prefers-contrast: high) {
+  .mobile-debugger {
+    background: #000;
+    border: 2px solid #fff;
+  }
+  
+  .log-entry {
+    border: 1px solid #fff;
+  }
+  
+  .debug-floating-button {
+    border: 2px solid #fff;
+  }
+}
+
+/* Reduced motion preferences */
+@media (prefers-reduced-motion: reduce) {
+  .debug-floating-button {
+    transition: none;
+  }
+  
+  .debug-floating-button:hover {
+    transform: none;
+  }
+  
+  .log-entry {
+    transition: none;
+  }
+}