--- a/frontend/src/components/mobile/index.ts
+++ b/frontend/src/components/mobile/index.ts
@@ -1,0 +1,18 @@
+// Mobile Components Index
+export { MobileOptimizedButton } from './MobileOptimizedButton';
+export { MobileOptimizedInput } from './MobileOptimizedInput';
+export { MobileSearchBar } from './MobileSearchBar';
+export { MobileNotesCard } from './MobileNotesCard';
+export { EnhancedMobileMenu } from './EnhancedMobileMenu';
+export { MobileLayout } from './MobileLayout';
+export { MobilePerformanceDashboard } from './MobilePerformanceDashboard';
+export { MobileDebugger } from './MobileDebugger';
+export { MobileTestSuite } from './MobileTestSuite';
+
+// Mobile Performance and Testing
+export { useMobilePerformance, performanceUtils } from '../../hooks/useMobilePerformance';
+
+// Re-export common types
+export type { MobileButtonProps } from './MobileOptimizedButton';
+export type { MobileInputProps } from './MobileOptimizedInput';
+export type { MobileMenuProps } from './EnhancedMobileMenu';