… MOBILE_PERFORMANCE_TESTING_GUIDE.md
… frontend/src/components/mobile/MobilePerformanceDashboard.css
… frontend/src/components/mobile/MobilePerformanceDashboard.tsx
… frontend/src/components/mobile/MobilePerformanceDemo.css
… frontend/src/components/mobile/MobilePerformanceDemo.tsx
… frontend/src/components/mobile/MobileTestSuite.css
… frontend/src/components/mobile/MobileTestSuite.tsx
… frontend/src/hooks/useMobilePerformance.ts